Fixing Loading Issue + Added Black Highlight
Just added a <script> tag right after header. Because HTML loads from
top to bottom, adding the <script> tag afterward, fixes it directly
afterward. The issue issue with having shouldResize()in the head tag is
that the header hasn’t event loaded so it’s invisible to the DOM (I
think). Whatever, it works now.