Terminal and Background Image
Updated the Terminal to have a highlight color of #0000ff.
Additionally, I fixed it’s scaling so that when the screen was smaller
than 615px, it would push some text to the next line. I also changed
the background image to a version without the text and team number. I
kind of like it with the text, but scaling was a bit odd because it
wasn’t symmetrical— I think it’s better this way.